+++ oap-server/oal-rt/src/main/java/org/apache/skywalking/oal/rt/parser/SourceColumnsFactory.java
@@ -32,7 +32,7 @@ public class SourceColumnsFactory {
             sourceColumns.add(
                 new SourceColumn(defaultColumn.getFieldName(), defaultColumn.getColumnName(),
                                  defaultColumn.getType(), defaultColumn.isID(), defaultColumn.getLength(),
-                                 defaultColumn.getIdxOfCompositeID(), defaultColumn.isGroupByCondInTopN()
+                                 defaultColumn.isCompositeID(), defaultColumn.getIdxOfCompositeID(), defaultColumn.isGroupByCondInTopN()
                 ));
         }
         return sourceColumns;